refactor: uculize navigation block
