+++ README.md
@@ -4,8 +4,8 @@ List of operations and their syntax:
 
 Navigation & working directory:
 
-- up - Go upper from current directory
 - cd path_to_directory - Go to dedicated folder from current directory
+- up - Go upper from current directory
 - ls - Print in console list of all files and folders in current directory
 
 Basic operations with files:
@@ -18,6 +18,7 @@ Basic operations with files:
 - rm path_to_file - Delete file
 
 Operating system info (prints following information in console)
+
 - os --EOL - Get EOL (default system End-Of-Line) and print it to console
 - os --cpus - Get host machine CPUs info (overall amount of CPUS plus model and clock rate (in GHz) for each of them) and print it to console
 - os --homedir - Get home directory and print it to console
@@ -25,8 +26,10 @@ Operating system info (prints following information in console)
 - os --architecture - Get CPU architecture for which Node.js binary has compiled and print it to console
 
 Hash calculation
- - hash path_to_file - Calculate hash for file and print it into console
+
+- hash path_to_file - Calculate hash for file and print it into console
 
 Compress and decompress operations
+
 - compress path_to_file path_to_destination - Compress file (using Brotli algorithm)
 - decompress path_to_file path_to_destination - Decompress file (using Brotli algorithm)
